cloud-functions/gemini-synthesis/requirements.txt
functions-framework==3.5.0
google-auth==2.23.0
requests==2.31.0